+++ sqlstore/src/main/resources/db/migration/V8502__seqparam_unique.sql
@@ -1,0 +1,2 @@
+ALTER TABLE SequencingParameters MODIFY COLUMN name varchar(255) NOT NULL;
+ALTER TABLE SequencingParameters ADD CONSTRAINT uk_sequencingParameters_name_model UNIQUE (name, instrumentModelId);